refactor: reorganize community module into vertical slices
