+++ BreastCancer.Tests/Unit/CommunityPostCreatedEventTests.cs
@@ -2,8 +2,10 @@ using AutoMapper;
 using BreastCancer.Community;
 using BreastCancer.Community.DTO.request;
 using BreastCancer.Community.DTO.response;
-using BreastCancer.Community.Events;
+using BreastCancer.Community.Events.Models;
+using BreastCancer.Community.Events.Sinks;
 using BreastCancer.Community.Features;
+using BreastCancer.Community.Features.CreatePost;
 using BreastCancer.Enum;
 using BreastCancer.Models;
 using BreastCancer.Repository.Interface;
