--- Community/Events/FanoutJob.cs
@@ -1,9 +1,0 @@
-using BreastCancer.Enum;
-
-namespace BreastCancer.Community.Events;
-
-public sealed record FanoutJob(
-    int PostId,
-    string AuthorId,
-    PostVisibility Visibility,
-    DateTimeOffset Timestamp);
