+++ Community/Events/Models/FollowCreatedEvent.cs
@@ -1,5 +1,3 @@
-using BreastCancer.Community.Domain;
-
-namespace BreastCancer.Community.Events;
+namespace BreastCancer.Community.Events.Models;
 
 public sealed record FollowCreatedEvent(string FollowerId, string FollowingId) : DomainEvent;
