+++ Community/Events/Models/PostCreatedEvent.cs
@@ -1,6 +1,5 @@
-using BreastCancer.Community.Domain;
 using BreastCancer.Enum;
 
-namespace BreastCancer.Community.Events;
+namespace BreastCancer.Community.Events.Models;
 
 public sealed record PostCreatedEvent(int PostId, string AuthorId, PostVisibility Visibility) : DomainEvent;
