+++ Community/Features/CreatePost/CreatePostCommand.cs
@@ -2,6 +2,6 @@ using BreastCancer.Community.DTO.request;
 using BreastCancer.Community.DTO.response;
 using MediatR;
 
-namespace BreastCancer.Community.Features;
+namespace BreastCancer.Community.Features.CreatePost;
 
 public sealed record CreatePostCommand(CreatePostDTO Post, string AuthorId, IReadOnlyCollection<string> Roles) : IRequest<PostDTO>;
