+++ Community/Features/FollowUser/FollowUserCommand.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 
-namespace BreastCancer.Community.Features
+namespace BreastCancer.Community.Features.FollowUser
 {
     public sealed record FollowUserCommand(string FollowerId, string FollowingId) : IRequest<Unit>;
 }
