+++ Community/Features/UnfollowUser/UnfollowUserCommand.cs
@@ -1,5 +1,5 @@
 using MediatR;
 
-namespace BreastCancer.Community.Features;
+namespace BreastCancer.Community.Features.UnfollowUser;
 
 public sealed record UnfollowUserCommand(string FollowerId, string FollowingId) : IRequest<Unit>;
